chore: Упразднил InfiniteDataStorage
