+++ package/src/InfiniteQuery/index.ts
@@ -1,3 +1,1 @@
 export * from './InfiniteQuery';
-
-export * from './InfiniteDataStorage';
