+++ package/src/Query/index.ts
@@ -1,3 +1,1 @@
 export * from './Query';
-
-export * from './DataStorage';
